feat: add GnuPG tool integration for system package signing
