+++ src/briefcase/integrations/__init__.py
@@ -5,6 +5,7 @@ from . import (
     file,
     flatpak,
     git,
+    gnupg,
     java,
     linuxdeploy,
     rcedit,
@@ -23,6 +24,7 @@ __all__ = [
     "file",
     "flatpak",
     "git",
+    "gnupg",
     "java",
     "linuxdeploy",
     "rcedit",
